Add files via upload
This is an HTML template for a Password Manager app with options to add and retrieve passwords. It includes a basic structure, buttons to toggle between forms, and inputs for account names and passwords. Styling is linked via style.css, and functionality is handled by script.js.



This is an HTML template for a Password Manager app with options to add and retrieve passwords. It includes a basic structure, buttons to toggle between forms, and inputs for account names and passwords. Styling is linked via style.css, and functionality is handled by script.js.